requirements.txt
httpx>=0.25.0
tiktoken>=0.5.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
